Add files via upload
The "Optimized Password Cracker" is a sophisticated tool designed in python so that I can demonstrate the process of password cracking using advanced hashing algorithms. This project leverages Python, Flask, and various hashing techniques such as SHA (Secure Hash Algorithm) and bcrypt to simulate and optimize the password cracking process. Users interact with the system through a web interface to upload a dump file containing example passwords and their corresponding hashes, which the application then attempts to crack.